parseEntityEntry now strips double quotes from spawnerTag values.  The vanilla JsonToNBT.func_150315_a() call to deserialize the NBT data from the spawnerTag JSON text does not function with property names wrapped in quotes.  That is technically illegal JSON, and the CoFH parser appears to attempt to rectify it.
Attempting to just wrap the spawnerTag in quotes turns the data into a
JSON string, which comes back as NBTTagString instead of an NBTTagCompound.

I was able to represent a wither skeleton spawner after making this change.

This is how I was sending expressing the spawnerTag property in a CoFH
generator JSON file:

"spawnEntity": {
	"spawnerTag": {EntityId:Skeleton,SpawnData:{SkeletonType:1}}
},