Add a hack for packets.
Basically. Tile packets changd in 1.10. getUpdateTag is a must and sent
with the initial chunk paket.
getUpdayePacket is the equevilent of getDescPacket.

PacketBases will now directly write to a packet buffer, then pipe
through an NBT tag, only for these special packet reqests. They are then
un wrapped from the tag, and piped through the rest of the packet pipe.