Gradle Buildscript.
CoFHLib is packed in built jars of Core.
AccessTransformer has been nuked. Add things back as they are needed. If
there is a getter / setter, use it OR submit a PR for a getter / setter
if it is usefull.
Avid using * in access transformers everywhere. Try and make them as
less invaisve as possible.