Fix an error when cloning methods for substitutable
Turns out we deleted the old one. Apparently.

Maybe this clone method works, maybe it doesn't.
The implementations of clone I looked at use Object and instanceof and
lookups through this map for some objects but just pass references to
others. It'll eihter work or crash, but we already crash so whatever
zero-sum worst-case.

Why was this so hard to implement anyway?
Why did  need to implement this?
Why didn't the library just provide this?